refactor: rename process_utils to processing_utils to match miles naming
